Remove legacy DefinesPlatformSettingsDrawer
