+++ Packages/UGF.Defines/Editor/DefinesPlatformSettingsDrawer.cs
@@ -1,7 +1,10 @@
-﻿using System;
+﻿using System.Collections.Generic;
 using UGF.EditorTools.Editor.Defines;
+using UGF.EditorTools.Editor.IMGUI;
 using UGF.EditorTools.Editor.IMGUI.EnabledProperty;
 using UGF.EditorTools.Editor.IMGUI.PlatformSettings;
+using UGF.EditorTools.Editor.IMGUI.Scopes;
+using UGF.EditorTools.Editor.Platforms;
 using UnityEditor;
 using UnityEngine;
 
@@ -9,46 +12,66 @@ namespace UGF.Defines.Editor
 {
     internal class DefinesPlatformSettingsDrawer : PlatformSettingsDrawer
     {
-        public event DefineGroupChangeHandler Applied;
-        public event DefineGroupChangeHandler Cleared;
-
-        public delegate void DefineGroupChangeHandler(string name, BuildTargetGroup buildTargetGroup);
+        private readonly Dictionary<string, ReorderableListDrawer> m_listDrawers = new Dictionary<string, ReorderableListDrawer>();
+        private const float PADDING = 5F;
 
         private Styles m_styles;
 
         private class Styles
         {
-            public GUIContent IncludeInBuild { get; } = new GUIContent("Include In Build", "Determines whether to include enabled defines in player build.");
-            public GUIContent Count { get; } = new GUIContent("Count");
-            public GUIContent ApplyContent { get; } = new GUIContent("Apply", "Apply specified defines to Project Settings compile symbols.");
-            public GUIContent ClearContent { get; } = new GUIContent("Clear", "Clear Project Settings compile symbols from all specified defines.");
             public GUIContent FlagOnContent { get; } = new GUIContent(EditorGUIUtility.FindTexture("Valid"), "Define currently included in compile symbols.");
             public GUIContent FlagOffContent { get; } = new GUIContent("X", "Define currently NOT included in compile symbols.");
 
             public GUIStyle FlagStyle { get; } = new GUIStyle("MiniLabel")
             {
                 alignment = TextAnchor.MiddleCenter,
-                normal =
-                {
-                    textColor = Color.grey
-                }
+                normal = { textColor = Color.grey }
             };
         }
 
         public DefinesPlatformSettingsDrawer()
         {
-            AutoSettingsInstanceCreation = true;
             AllowEmptySettings = false;
+            AutoSettingsInstanceCreation = true;
         }
 
-        public void SetupGroupTypes()
+        public string GetSelectedPlatformName()
         {
-            for (int i = 0; i < Groups.Count; i++)
-            {
-                string group = Groups[i];
+            return GetSelectedGroupName();
+        }
 
-                AddGroupType(group, typeof(DefinesSettings));
+        protected override void OnEnable()
+        {
+            base.OnEnable();
+
+            AddPlatformAllAvailable();
+
+            for (int i = 0; i < PlatformEditorUtility.PlatformsAllAvailable.Count; i++)
+            {
+                PlatformInfo platform = PlatformEditorUtility.PlatformsAllAvailable[i];
+
+                AddGroupType(platform.Name, typeof(DefinesSettings));
             }
+
+            foreach (KeyValuePair<string, ReorderableListDrawer> pair in m_listDrawers)
+            {
+                pair.Value.Enable();
+            }
+        }
+
+        protected override void OnDisable()
+        {
+            base.OnDisable();
+
+            ClearGroups();
+            ClearGroupTypes();
+
+            foreach (KeyValuePair<string, ReorderableListDrawer> pair in m_listDrawers)
+            {
+                pair.Value.Disable();
+            }
+
+            m_listDrawers.Clear();
         }
 
         protected override void OnDrawGUI(Rect position, SerializedProperty propertyGroups)
@@ -60,208 +83,85 @@ namespace UGF.Defines.Editor
 
         protected override void OnDrawSettings(Rect position, SerializedProperty propertyGroups, string name)
         {
+            float height = EditorGUIUtility.singleLineHeight;
             float space = EditorGUIUtility.standardVerticalSpacing;
 
-            float propertiesHeight = OnGetPropertiesHeight(propertyGroups);
-            float elementsHeight = OnGetElementsHeight(propertyGroups);
-            float controlsHeight = OnGetControlsHeight(propertyGroups);
-
-            var propertiesPosition = new Rect(position.x, position.y, position.width, propertiesHeight);
-            var elementsPosition = new Rect(position.x, propertiesPosition.yMax + space, position.width, elementsHeight);
-            var controlsPosition = new Rect(position.x, elementsPosition.yMax + space, position.width, controlsHeight);
-
-            OnDrawProperties(propertiesPosition, propertyGroups, name);
-            OnDrawElements(elementsPosition, propertyGroups, name);
-            OnDrawControls(controlsPosition, propertyGroups, name);
-        }
-
-        protected virtual void OnDrawProperties(Rect position, SerializedProperty propertyGroups, string name)
-        {
             SerializedProperty propertySettings = OnGetSettings(propertyGroups, name);
-            SerializedProperty propertyDefines = propertySettings.FindPropertyRelative("m_defines");
-            SerializedProperty propertyIncludeInBuild = propertySettings.FindPropertyRelative("m_includeInBuild");
-            SerializedProperty propertySize = propertyDefines.FindPropertyRelative("Array.size");
+            ReorderableListDrawer drawer = GetListDrawer(name, propertySettings);
 
-            float line = EditorGUIUtility.singleLineHeight;
-            float space = EditorGUIUtility.standardVerticalSpacing;
-
-            var includeInBuildPosition = new Rect(position.x, position.y, position.width, line);
-            var sizePosition = new Rect(position.x, includeInBuildPosition.yMax + space, position.width, line);
-
-            EditorGUI.PropertyField(includeInBuildPosition, propertyIncludeInBuild, m_styles.IncludeInBuild);
-            EditorGUI.PropertyField(sizePosition, propertySize, m_styles.Count);
-        }
-
-        protected virtual void OnDrawElements(Rect position, SerializedProperty propertyGroups, string name)
-        {
-            SerializedProperty propertySettings = OnGetSettings(propertyGroups, name);
-            SerializedProperty propertyDefines = propertySettings.FindPropertyRelative("m_defines");
-
-            float space = EditorGUIUtility.standardVerticalSpacing;
-
-            for (int i = 0; i < propertyDefines.arraySize; i++)
+            using (new LabelWidthChangeScope(-PADDING))
             {
-                float height = OnGetElementHeight(propertyGroups, i);
+                var rectPlatformName = new Rect(position.x, position.y, position.width, height);
+                var rectIncludeInBuild = new Rect(position.x, rectPlatformName.yMax + space, position.width, height);
+                var rectDefines = new Rect(position.x, rectIncludeInBuild.yMax + space, position.width, height);
 
-                position.height = height;
+                SerializedProperty propertyIncludeInBuild = propertySettings.FindPropertyRelative("m_includeInBuild");
 
-                OnDrawElement(position, propertyGroups, i);
+                OnDrawSettingsPlatformName(rectPlatformName, propertyGroups, name);
 
-                position.y += height + space;
-            }
-        }
+                EditorGUI.PropertyField(rectIncludeInBuild, propertyIncludeInBuild);
 
-        protected virtual void OnDrawElement(Rect position, SerializedProperty propertyGroups, int index)
-        {
-            string name = GetSelectedGroupName();
-            SerializedProperty propertySettings = OnGetSettings(propertyGroups, name);
-            SerializedProperty propertyDefines = propertySettings.FindPropertyRelative("m_defines");
-            SerializedProperty propertyElement = propertyDefines.GetArrayElementAtIndex(index);
-
-            float line = EditorGUIUtility.singleLineHeight;
-            float space = EditorGUIUtility.standardVerticalSpacing;
-
-            var propertyPosition = new Rect(position.x, position.y, position.width - line - space, position.height);
-            var flagPosition = new Rect(propertyPosition.xMax + space, position.y, line, position.height);
-
-            EnabledPropertyGUIUtility.EnabledProperty(propertyPosition, GUIContent.none, propertyElement);
-
-            DrawDefineFlag(flagPosition, propertyGroups, index);
-        }
-
-        protected virtual void OnDrawControls(Rect position, SerializedProperty propertyGroups, string name)
-        {
-            SerializedProperty propertySettings = OnGetSettings(propertyGroups, name);
-            SerializedProperty propertyDefines = propertySettings.FindPropertyRelative("m_defines");
-
-            float line = EditorGUIUtility.singleLineHeight;
-            float space = EditorGUIUtility.standardVerticalSpacing;
-            float width = 50F;
-
-            position.y += space;
-
-            bool hasApply = Applied != null;
-            bool hasClear = Cleared != null;
-            bool hasAny = hasApply || hasClear;
-            bool hasDefines = propertyDefines.arraySize > 0;
-
-            var applyPosition = new Rect(position.xMax - width - line, position.y, width, line);
-            var clearPosition = new Rect(applyPosition.x - width - space, position.y, width, line);
-
-            if (hasAny)
-            {
-                using (new EditorGUI.DisabledScope(!hasApply || !hasDefines))
-                {
-                    if (GUI.Button(applyPosition, m_styles.ApplyContent))
-                    {
-                        OnApply(propertyGroups, name);
-                    }
-                }
-
-                using (new EditorGUI.DisabledScope(!hasClear || !hasDefines))
-                {
-                    if (GUI.Button(clearPosition, m_styles.ClearContent))
-                    {
-                        OnClearAll(propertyGroups, name);
-                    }
-                }
+                drawer.DrawGUI(rectDefines);
             }
         }
 
         protected override float OnGetSettingsHeight(SerializedProperty propertyGroups)
         {
+            float height = EditorGUIUtility.singleLineHeight;
             float space = EditorGUIUtility.standardVerticalSpacing;
-            float propertiesHeight = OnGetPropertiesHeight(propertyGroups);
-            float elementsHeight = OnGetElementsHeight(propertyGroups);
-            float controlsHeight = OnGetControlsHeight(propertyGroups);
 
-            if (controlsHeight > 0F)
-            {
-                controlsHeight += 5F;
-            }
-            else
-            {
-                controlsHeight = space;
-            }
-
-            return propertiesHeight + elementsHeight + controlsHeight + space * 4F;
-        }
-
-        protected virtual float OnGetPropertiesHeight(SerializedProperty propertyGroups)
-        {
-            return EditorGUIUtility.singleLineHeight * 2F + EditorGUIUtility.standardVerticalSpacing;
-        }
-
-        protected virtual float OnGetElementsHeight(SerializedProperty propertyGroups)
-        {
             string name = GetSelectedGroupName();
             SerializedProperty propertySettings = OnGetSettings(propertyGroups, name);
-            SerializedProperty propertyDefines = propertySettings.FindPropertyRelative("m_defines");
+            ReorderableListDrawer drawer = GetListDrawer(name, propertySettings);
 
+            float heightDefines = drawer.SerializedProperty.isExpanded ? drawer.GetHeight() : height;
+
+            return height * 2F + space * 3F + heightDefines + PADDING * 2F;
+        }
+
+        protected override void OnCreateSettings(SerializedProperty propertyGroups, string name, SerializedProperty propertySettings)
+        {
+            base.OnCreateSettings(propertyGroups, name, propertySettings);
+
+            GetListDrawer(name, propertySettings);
+        }
+
+        private ReorderableListDrawer GetListDrawer(string name, SerializedProperty propertySettings)
+        {
+            if (!m_listDrawers.TryGetValue(name, out ReorderableListDrawer drawer))
+            {
+                SerializedProperty propertyDefines = propertySettings.FindPropertyRelative("m_defines");
+
+                drawer = new ReorderableListDrawer(propertyDefines);
+                drawer.List.drawElementCallback = (rect, index, active, focused) => { OnListDrawElement(drawer, name, rect, index); };
+
+                m_listDrawers.Add(name, drawer);
+            }
+
+            return drawer;
+        }
+
+        private void OnListDrawElement(ReorderableListDrawer drawer, string name, Rect rect, int index)
+        {
+            PlatformInfo platform = PlatformEditorUtility.GetPlatform(name);
+
+            float height = EditorGUIUtility.singleLineHeight;
             float space = EditorGUIUtility.standardVerticalSpacing;
-            float height = 0F;
 
-            for (int i = 0; i < propertyDefines.arraySize; i++)
-            {
-                height += OnGetElementHeight(propertyDefines, i) + space;
-            }
+            rect.y += space;
 
-            return height;
-        }
-
-        protected virtual float OnGetElementHeight(SerializedProperty propertyGroups, int index)
-        {
-            return EditorGUIUtility.singleLineHeight;
-        }
-
-        protected virtual float OnGetControlsHeight(SerializedProperty propertyGroups)
-        {
-            float line = EditorGUIUtility.singleLineHeight;
-            float space = EditorGUIUtility.standardVerticalSpacing;
-
-            bool hasApply = Applied != null;
-            bool hasClear = Cleared != null;
-
-            return hasApply || hasClear ? line + space * 2F : 0F;
-        }
-
-        protected virtual void OnApply(SerializedProperty propertyGroups, string name)
-        {
-            if (Enum.TryParse(name, out BuildTargetGroup group))
-            {
-                Applied?.Invoke(name, group);
-            }
-        }
-
-        protected virtual void OnClearAll(SerializedProperty propertyGroups, string name)
-        {
-            if (Enum.TryParse(name, out BuildTargetGroup group))
-            {
-                Cleared?.Invoke(name, group);
-            }
-        }
-
-        protected virtual void DrawDefineFlag(Rect position, SerializedProperty propertyGroups, int index)
-        {
-            string name = GetSelectedGroupName();
-            SerializedProperty propertySettings = OnGetSettings(propertyGroups, name);
-            SerializedProperty propertyDefines = propertySettings.FindPropertyRelative("m_defines");
-            SerializedProperty propertyElement = propertyDefines.GetArrayElementAtIndex(index);
+            SerializedProperty propertyElement = drawer.SerializedProperty.GetArrayElementAtIndex(index);
             SerializedProperty propertyValue = propertyElement.FindPropertyRelative("m_value");
 
-            GUIContent flagContent = m_styles.FlagOffContent;
+            var rectDefine = new Rect(rect.x, rect.y, rect.width - height + space, height);
+            var rectFlag = new Rect(rectDefine.xMax + space, rect.y, height, height);
 
-            if (!string.IsNullOrEmpty(propertyValue.stringValue) && Enum.TryParse(name, out BuildTargetGroup value))
-            {
-                bool hasDefine = DefinesEditorUtility.HasDefine(propertyValue.stringValue, value);
+            GUIContent flagContent = !string.IsNullOrEmpty(propertyValue.stringValue) && DefinesEditorUtility.HasDefine(propertyValue.stringValue, platform.BuildTargetGroup)
+                ? m_styles.FlagOnContent
+                : m_styles.FlagOffContent;
 
-                if (hasDefine)
-                {
-                    flagContent = m_styles.FlagOnContent;
-                }
-            }
-
-            GUI.Label(position, flagContent, m_styles.FlagStyle);
+            EnabledPropertyGUIUtility.EnabledProperty(rectDefine, GUIContent.none, propertyElement, false);
+            GUI.Label(rectFlag, flagContent, m_styles.FlagStyle);
         }
     }
 }
