--- Packages/UGF.Defines/Editor/DefinesPlatformSettingsDrawer2.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 54afdc0baedf462a8159ac8587aaa494
-timeCreated: 1630764383
